+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * `assertIsArray()`, `assertIsInt()`, `assertIsObject()` and `assertIsString()`
     methods of `TestTrait` are now provided by `__call()` and `__callStatic()`
     methods. These also provide some other "assertIs" methods (see API);
+* fixed bug for `assertException()` assert method, it checks if the
+    `$expectedException` is a subclass of `Exception`;
 * fixed bug for `assertIsArrayNotEmpty()` assert method, it executes
     `array_filter()` on the array to verify that it does not contain a value
     that is nevertheless equal to empty;